chore: move catalogModuleGithubOrgEntityProvider to its own backend module
