+++ .changeset/strange-dodos-film.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-backend-module-github-org': minor
+---
+
+Added `catalogModuleGithubOrgEntityProvider` to ingest users and teams from multiple Github organizations.
