+++ plugins/catalog-backend-module-github-org/README.md
@@ -1,0 +1,5 @@
+# @backstage/plugin-catalog-backend-module-github-org
+
+The github-org backend module for the catalog plugin.
+
+_This plugin was created through the Backstage CLI_
