+++ plugins/catalog-backend-module-github-org/api-report.md
@@ -1,0 +1,11 @@
+## API Report File for "@backstage/plugin-catalog-backend-module-github-org"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { BackendFeature } from '@backstage/backend-plugin-api';
+
+// @public
+const catalogModuleGithubOrgEntityProvider: () => BackendFeature;
+export default catalogModuleGithubOrgEntityProvider;
+```
