+++ plugins/catalog-backend-module-github-org/src/module.test.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,7 +19,7 @@ import { mockServices, startTestBackend } from '@backstage/backend-test-utils';
 import { EntityProvider } from '@backstage/plugin-catalog-node';
 import { catalogProcessingExtensionPoint } from '@backstage/plugin-catalog-node/alpha';
 import { Duration } from 'luxon';
-import { catalogModuleGithubOrgEntityProvider } from './catalogModuleGithubOrgEntityProvider';
+import { catalogModuleGithubOrgEntityProvider } from './module';
 
 describe('catalogModuleGithubOrgEntityProvider', () => {
   it('should register provider at the catalog extension point', async () => {
