+++ plugins/catalog-backend-module-github-org/src/module.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Backstage Authors
+ * Copyright 2023 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -64,7 +64,7 @@ export const githubOrgEntityProviderTransformsExtensionPoint =
 /**
  * Registers the `GithubMultiOrgEntityProvider` with the catalog processing extension point.
  *
- * @alpha
+ * @public
  */
 export const catalogModuleGithubOrgEntityProvider = createBackendModule({
   pluginId: 'catalog',
